+++ tests/testthat/test-vec-new.R
@@ -112,3 +112,8 @@ test_that("rep_named() repeats along names", {
 test_that("rep_along() reps along vector", {
   expect_identical(rep_along(1:2, list(zap())), list(zap(), zap()))
 })
+
+test_that("chr() supports logical NA", {
+  expect_identical(chr(NA), na_chr)
+  expect_identical(chr(NA, NA), c(na_chr, na_chr))
+})